art_snob_primrose/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
alabaster==0.7.12
annoy==1.16.3
appnope==0.1.0
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
atlas==0.27.0
attrs==19.3.0
aws-sam-translator==1.23.0
aws-xray-sdk==2.5.0
Babel==2.8.0
backcall==0.1.0
backoff==1.10.0
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
cfn-lint==0.31.1
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
coverage==5.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
docker==4.2.0
docutils==0.16
ecdsa==0.15
future==0.18.2
gast==0.3.3
gcloud-aio-auth==3.1.0
gcloud-aio-storage==5.4.0
google-api-core==1.17.0
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-logging==1.15.0
google-cloud-datastore==1.12.0
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
h5py==2.10.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.6.0
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsondiff==1.1.2
jsonpatch==1.25
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
jstyleson==0.0.2
junit-xml==1.9
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.32.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mock==4.0.2
more-itertools==8.3.0
moto==1.3.14
multidict==4.7.6
mysql-connector-python==8.0.20
networkx==2.4
nltk==3.5
numba==0.49.1
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pockets==0.9.1
primrose==1.0.10
prompt-toolkit==3.0.5
protobuf==3.12.0
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytest-cov==2.8.1
python-dateutil==2.8.0
python-jose==3.1.0
pytz==2020.1
PyYAML==5.3.1
recommonmark==0.6.0
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.14
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
slackclient==2.4.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sshpubkeys==3.1.0
tbb==2019.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
testfixtures==6.14.1
threadpool==1.3.2
threadpoolctl==2.0.0
tqdm==4.46.0
traitlets==4.3.3
typing==3.7.4.1
umap-learn==0.4.3
urllib3==1.25.9
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.4.2
zipp==3.1.0
